6 Size of@

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ long int b;
 long long int c;
 float f;
 
-printf("size of a char: %ld byte(s)\n", sizeof(d));
-printf("size of an int: %ld byte(s)\n", sizeof(a));
-printf("size of a long int: %ld byte(s)\n", sizeof(b));
-printf("size of a long long int: %ld byte(s)\n", sizeof(c));
-printf("size of a float: %ld byte(s)\n", sizeof(f));
+printf("size of a char: %ld byte(s)\n", (unsigned long)sizeof(d));
+printf("size of an int: %ld byte(s)\n", (unsigned long)sizeof(a));
+printf("size of a long int: %ld byte(s)\n", (unsigned long)sizeof(b));
+printf("size of a long long int: %ld byte(s)\n", (unsigned long)sizeof(c));
+printf("size of a float: %ld byte(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
